feat: replace and delete (#1)
This commit adds replace and delete methods so cache managers can do
both operations.

Replace allows you to imperatively change what's stored in the cache for
a given key. This use case is very common when you are listening to
events that need to actively update the cache.

Delete allows you to remove data from the cache actively. The use case
here is when you want to invalidate data from the cache.